Create CommentCreateMixin to avoid django and drf code duplication
