+++ api/v1/interactions/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path, include
+from django.urls import include, path
 
 app_name = "interactions"
